redesigned reset circuit to use MAX708
This avoids the need to generate pulses for the
watchdog timer, which we don't want or need.
Also, eliminated the '393 counter to divide the
clock, since A19 (which will go high the first time
the code jumps to an address at 80000 or above)
can serve as the trigger to disable the ROM
aliasing.

The fewer chips, the better!